Add project listing feature and unify ID types
Introduces a new Project entity, repository, and controller on the backend to support project listings. Updates the Personal page to display projects dynamically, adds related frontend components, and includes a resume PDF. Refactors backend and API code to use Integer instead of Long for entity IDs, improving consistency. Also adds a Dockerfile for the client, a new Footer component, and various UI improvements.

+++ server/src/main/java/wiki/ednotes/server/article/Article.java
@@ -7,7 +7,7 @@ import jakarta.persistence.*;
 public class Article {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "title", nullable = false)
     private String title;
@@ -16,7 +16,7 @@ public class Article {
     private String content;
 
     @Column(name = "category_id")
-    private Long categoryId;
+    private Integer categoryId;
 
     @Column(name = "is_published")
     private Boolean isPublished;
@@ -24,11 +24,11 @@ public class Article {
     @Column(name = "order_in_category")
     private Integer orderInCategory;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -48,11 +48,11 @@ public class Article {
         this.content = content;
     }
 
-    public Long getCategoryId() {
+    public Integer getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(Long categoryId) {
+    public void setCategoryId(Integer categoryId) {
         this.categoryId = categoryId;
     }
